+++ packages/slate-react/src/react-reconciler.d.ts
@@ -1,0 +1,4 @@
+declare module "react-reconciler" {
+  const Reconciler: (hostConfig: unknown) => any;
+  export default Reconciler;
+}